+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.bignerdranch.android.koala.KoalaFieldInject
 import kotlinx.coroutines.launch
-import woowacourse.shopping.annotation.DataBaseCartRepository
+import woowacourse.shopping.di.annotation.DataBaseCartRepository
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
